Fix keyboard shortcuts to use Control instead of Meta on Windows and Linux
