+++ src/lib/utils.ts
@@ -1,6 +1,7 @@
 import { SourceRange } from '../lang/wasm'
 
 import { v4 } from 'uuid'
+import { isDesktop } from './isDesktop'
 
 export const uuidv4 = v4
 
@@ -124,6 +125,41 @@ export function getNormalisedCoordinates({
     x: Math.round((browserX / width) * streamWidth),
     y: Math.round((browserY / height) * streamHeight),
   }
+}
+
+// TODO: Remove the empty platform type.
+export type Platform = 'macos' | 'windows' | 'linux' | ''
+
+export function platform(): Platform {
+  if (isDesktop()) {
+    const platform = window.electron.platform ?? ''
+    // https://nodejs.org/api/process.html#processplatform
+    switch (platform) {
+      case 'darwin':
+        return 'macos'
+      case 'win32':
+        return 'windows'
+      // We don't currently care to distinguish between these.
+      case 'android':
+      case 'freebsd':
+      case 'linux':
+      case 'openbsd':
+      case 'sunos':
+        return 'linux'
+      default:
+        console.error('Unknown platform:', platform)
+        return ''
+    }
+  }
+  if (navigator.userAgent.indexOf('Mac') !== -1) {
+    return 'macos'
+  } else if (navigator.userAgent.indexOf('Win') !== -1) {
+    return 'windows'
+  } else if (navigator.userAgent.indexOf('Linux') !== -1) {
+    return 'linux'
+  }
+  console.error('Unknown platform userAgent:', navigator.userAgent)
+  return ''
 }
 
 export function isReducedMotion(): boolean {
